requirements.txt
Django==4.2.7
djangorestframework==3.14.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.0.3
joblib==1.3.2
